Allow reset of default backlog for server sockets
This commit is about three following changes:

1. Control of default backlog throughout the lifetime of a Tornado
application, through the now exported (no `_` prefix) module variable.
Previously, by way of how Python initializes function argument defaults,
the default was used once during module initialization (when procedures are
created) which "froze" the default value in the sense there was no way to
affect the default during application lifetime -- typically as
configuration is loaded, for example. To allow application to configure
the default backlog when it has the preferred value available, evaluation
of the default backlog immediately before calling `listen`, is thus
necessary.

2. Deferring to Python deciding on the backlog, through using `None` as
valid value. For Python 3.6, for example, that would imply the minimum
between 128 and SOMAXCONN:
https://github.com/python/cpython/blob/3.6/Modules/socketmodule.c#L3026

3. Because the number of connections Tornado attempted to `accept` with
each iteration of it's I/O loop, was tied to the default backlog value,
because this commit allows for `None` or zero (valid for `listen`) values
for the latter, the heuristics needed to be amended because neither zero
nor `None` would be an acceptable number of connections to accept. This
problem is solved by appropriately adding a separate variable that
controls the number, with fallback through a non-zero default backlog to
128.